media: ov5640: Improve firmware load time
Downloading the firmware can be done in groups to minimize i2c packets.
The firmware status is set practically immediately upon successfully
loading it, so there's no need for a long timeout or retry. The
firmware can also not be loaded when the sensor isn't powered on, so
don't bother trying.