usb: dwc3: Add support for snps,usb3-phy-reset-quirk
RK3399 TypeC PHY needs to be powered off and powered on again
for it to apply the correct Type-C plug orientation setting from
extcon and reconfigure itself while the USB controller is held
in reset. (It can not just reconfigure itself without USB controller
driver cooperation due to this requirement.)

Good place to perform the power cycle is in __dwc3_set_mode
when changing between device and host modes. The only problem
is that __dwc3_set_mode will not get called in case the port
stays in device mode between plugout/plugin cycle into the same
type of the USB host port. DWC3 will not see a change in dr_role
but the user may have changed the orientation of the Type-C plug,
so the PHY may need a power cycle, which we'd like to perform in
__dwc3_set_mode.

We can make __dwc3_set_mode be called for plugout/plugin
events, when detected if we add a special value for dr_role
(DWC3_GCTL_PRTCAP_DEVICE_DISCONNECTED) that would express
the meaning of "nothing connected to the port".

For that purpose we observe complete USB disconnect via
lack of extcon USB and USB_HOST connector types in drd.c
and pass this state as DWC3_GCTL_PRTCAP_DEVICE_DISCONNECTED
to dwc3_set_mode().

It's a bit unfortunate that dr_role contains a direct register
value, so any value we add will not be a real register value,
but such is life.

Signed-off-by: Ondrej Jirman <megi@xff.cz>