media: sun6i-csi: add V4L2_CAP_IO_MC capability
Advertise the V4L2_CAP_IO_MC flag and update
sun6i_csi_capture_enum_fmt to work with this capability.

With the capability adverstised, the VIDIOC_ENUM_FMT ioctl
implementation is updated to enumerate pixel formats depending
on the supplied mbus_code.  If the code is 0, all formats are
enumerated, if not, only pixel formats that match with the supplied
mbus_code are enumeratd.

Signed-off-by: Adam Pigg <adam@piggz.co.uk>