usb: typec: anx7688: Don't use I2C regulator
anx7688 never disables any regulators across suspend. So we don't want
to keep this one enabled, beceuase on Pinephone it's also powering
touch controller which is hard to disable during sleep, so it needs
its power removed.

(Alternative here would be to disable I2C bus regulator in this driver
during suspend, since it's not needed for I2C comm during suspend. But
just not using it is easier for now.)

Signed-off-by: Ondrej Jirman <megi@xff.cz>