rtw88: Scaffolding for 8723cs (8703b chip) support
IF YOU RUN THIS, IT WILL CRASH.

Scaffolding means I've put the infrastructure in place that the SDIO
device should be matched to the driver. The functions marked as
"required" in rte8703b_ops are ones the shared rtw88 code calls
without checking if they exist, so it would most likely crash with a
NULL pointer dereference until those are implemented (or at least the
ones used during init, if they return error codes that stop the driver
from trying others).

Signed-off-by: Fiona Klute <fiona.klute@gmx.de>