Fix EVM dbm calculation
The value provided by the hardware is supposed to be a signed value
that's usually negative, but pkt_stat->rx_evm is an u8 array. This
means the value must be cast to s8 before the shift and sign flip
operation, or the clamp_t result will pretty much always be
0. Probably a bug in rtw8723d too, the other rtw88 chip drivers cast
the value before using. Maybe pkt_stat->rx_evm should just be an s8
array instead?

Signed-off-by: Fiona Klute <fiona.klute@gmx.de>